split iptables
